+++ src/components/Layout/DesktopGrid.module.css
@@ -64,7 +64,7 @@
   color: #000000;
   text-decoration: none;
   font-size: 2.2rem;
-  z-index: 1;
+  z-index: 2;
 
   &:hover {
     text-decoration: line-through;
