+++ src/components/Layout/MobileWrapper.tsx
@@ -5,6 +5,9 @@ import { Language } from "../../types";
 import { useGlobalBackground } from "../../state/global";
 import { backgroundCSS } from "./backgroundCSS";
 import { useLinks } from "../../hooks/contentful";
+import classNames from "classnames";
+import { useIsHome } from "./useIsHome";
+import { useIntersection } from "./useIntersection";
 
 interface Props extends PropsWithChildren {
   lang: Language;
@@ -15,12 +18,25 @@ export function MobileWrapper({ lang, children }: Props) {
     data: { items },
   } = useLinks();
 
-  const location = useLocation();
-  // eslint-disable-next-line
-  const isHome = location.pathname.match(/^\/(pl|en)[\/]?$/);
+  const { ref: intersectionRef, intersects } = useIntersection();
+  const isHome = useIsHome();
 
   return (
     <div style={backgroundCSS(background)} className={styles.wrapper}>
+      <div className={styles.header}>
+        <h1
+          className={classNames(styles.logo, { [styles.hidden]: intersects })}
+        >
+          problonde
+        </h1>
+      </div>
+      {isHome && (
+        <div className={styles.logoWrapper}>
+          <h1 className={styles.hugeLogo} ref={intersectionRef}>
+            problonde
+          </h1>
+        </div>
+      )}
       {children}
     </div>
   );
